test: move fixtures to `interchange.conftest`
